Added funcionalities: language selector, filter by category and filter by subcategory.
Added a combobox with languages and you can now load data in the selected language.
Implemented the category filter, and once you choose one, the subcategory items are loaded and the data too.
Implemented the subcategory filter, and once you choose on, the data with the subcategory selected is loaded.